Final version
In information theory, the Hamming distance between two sequences of equal length is the number of
positions at which the corresponding symbols are different. Put another way, it measures the minimum
number of substitutions required to change one string into the other, or the number of errors that
transform one string into the other.

Write a program that prompts the user to enter two numbers between 0-100 (decimal) and computes
the Hamming distance between the two numbers when the numbers are represented in base-3 (ternary
format). The program then displays the Hamming distance on the screen.