Final version
Write code that asks the user to enter a single letter. Then choose one of three operations: (1) change
case; (2) reverse alphabet; (3) encrypt.

1. Change case should convert the letter to the opposite case (upper to lower; lower to upper).
2. Reverse alphabet should find the corresponding letter starting from the end of the alphabet AND
change the case. For example, ‘a’ equals ‘Z’ and ‘Y’ equals ‘b’.
3. Encrypt should take a letter and change it six letters “ahead” (assume the alphabet “wraps around”
and all of the same case letters are together). For example ‘C’ is ‘I’ and ‘z’ is ‘F’

If the input is not a letter (e.g., user entered a number instead of a letter), your code should print an
error message. If the input is more than one letter, an error message should also be produced.